Refactor Beacon Process
