+++ mempool/shard_to_beacon_pool_v2.go
@@ -2,7 +2,6 @@ package mempool
 
 import (
 	"errors"
-	"fmt"
 	"sort"
 	"strconv"
 	"sync"
@@ -231,15 +230,13 @@ func (self *ShardToBeaconPool) GetValidBlock(limit map[byte]uint64) map[byte][]*
 			finalBlocks[shardID] = append(finalBlocks[shardID], blk)
 		}
 	}
-	//UNCOMMENT FOR TESTING
-	// Logger.log.Info()
-	// fmt.Print("ShardToBeaconPool/ValidPendingBlock ")
-	// for _, block := range finalBlocks[byte(0)] {
-	// 	fmt.Printf(" %+v ", block.Header.Height)
-	// }
-	// Logger.log.Info()
-	//==============
-	fmt.Println("GetValidBlock", limit, finalBlocks)
+	////UNCOMMENT FOR TESTING
+	//Logger.log.Infof("ShardToBeaconPool, Valid Block ")
+	//for _, block := range finalBlocks[byte(0)] {
+	//	fmt.Printf(" %+v ", block.Header.Height)
+	//}
+	////==============
+	//fmt.Println("GetValidBlock", limit, finalBlocks)
 	return finalBlocks
 }
